+++ readability/scripts/client.py
@@ -4,9 +4,9 @@
 A fast python port of arc90's readability tool
 
 Usage:
-    breadability [options] <resource>
-    breadability --version
-    breadability --help
+    readability [options] <resource>
+    readability --version
+    readability --help
 
 Arguments:
   <resource>      URL or file path to process in readable form.
@@ -42,7 +42,7 @@ def parse_args():
 
 def main():
     args = parse_args()
-    logger = logging.getLogger("breadability")
+    logger = logging.getLogger("readability")
 
     if args["--verbose"]:
         logger.setLevel(logging.DEBUG)